refactor(dashboard): extract macOS menu-bar wiring into useTauriMenuWiring
Move the six menu callbacks (#4695/#4942) and the useTauriMenuEvents
binding out of App.tsx into a dedicated hook. Callbacks are byte-identical;
no-op outside Tauri. App now passes the five inputs the wiring needs.

Part of #5560 (App.tsx Shell decomposition).
